[#96] Image-related namespaces (Part 39: Each 'CSharedPodArray' derivative can "grow", not just 'CSharedBytes')

+++ Main/src/Dos_FatPath.cpp
@@ -25,7 +25,7 @@
 
 	CDos::CFatPath::CFatPath(DWORD nItemsMax)
 		// ctor for Dummy object which has no Buffer and just counts the Items (allocation units)
-		: buffer(0)
+		: buffer((DWORD)0)
 		, nItems(0) , pLastItem(nullptr) , error(TError::OK) {
 		buffer.length=nItemsMax;
 	}
